requirements.txt
# Core dependencies
python-telegram-bot==20.7
sqlalchemy==2.0.23
alembic==1.13.1

# Data processing
pandas==2.1.4
numpy==1.26.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Financial data
yfinance==0.2.33
alpha-vantage==2.3.1

# Technical analysis
ta==0.11.0
ta-lib==0.4.28

# AI/ML (optional)
openai==1.6.1
scikit-learn==1.3.2

# Database
psycopg2-binary==2.9.9  # PostgreSQL
redis==5.0.1            # Cache

# Environment and config
python-dotenv==1.0.0
pydantic==2.5.3

# Async and scheduling
celery==5.3.4
apscheduler==3.10.4

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
babel==2.14.0

# Development
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
black==23.12.1
flake8==7.0.0
mypy==1.8.0
ipython==8.19.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Security
cryptography==41.0.7

# Logging and monitoring
loguru==0.7.2
sentry-sdk==1.39.1